refactor(actions): refactor cave actions
